refactor(deploy): Structure deploy into sub-packages and split graph deployment fully from old sequential deploy
Restructure deploy pkg into internal subpackages and modify some exported method names for readability.
This allows to split off the old sequential non-graph deployment into a subpackage without introducing any weird
imports or cycles. This new package will be possible to drop completely once graph based deploy becomes always active default.

Additionally the method to resolve parameter values wich is used in deploy, integration test asserts and downloads is moved to
config/parameter, where a method to resolve parameter values makes more sense anyway.
This method could possible be further refactored in the future and e.g. become a method of Config. For now, it is just moved
out of deploy.
