+++ pkg/deploy/internal/automation/automation.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package deploy
+package automation
 
 import (
 	"context"
@@ -25,23 +25,28 @@ import (
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/client/automation"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config"
 	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/config/parameter"
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/deploy/errors"
+	"github.com/dynatrace/dynatrace-configuration-as-code/v2/pkg/deploy/internal/resolve"
 )
 
-//go:generate mockgen -source=automation.go -destination=automation_mock.go -package=deploy automationClient
-type automationClient interface {
+//go:generate mockgen -source=automation.go -destination=automation_mock.go -package=automation automationClient
+type Client interface {
 	Upsert(ctx context.Context, resourceType automation.ResourceType, id string, data []byte) (result *automation.Response, err error)
 }
-type dummyAutomationClient struct {
+
+var _ Client = (*DummyClient)(nil)
+
+type DummyClient struct {
 }
 
-func (c *dummyAutomationClient) Upsert(_ context.Context, _ automation.ResourceType, id string, _ []byte) (*automation.Response, error) {
+func (c *DummyClient) Upsert(_ context.Context, _ automation.ResourceType, id string, _ []byte) (*automation.Response, error) {
 	return &automation.Response{ID: id}, nil
 }
 
-func deployAutomation(ctx context.Context, client automationClient, properties parameter.Properties, renderedConfig string, c *config.Config) (ResolvedEntity, error) {
+func Deploy(ctx context.Context, client Client, properties parameter.Properties, renderedConfig string, c *config.Config) (config.ResolvedEntity, error) {
 	t, ok := c.Type.(config.AutomationType)
 	if !ok {
-		return ResolvedEntity{}, newConfigDeployErr(c, fmt.Sprintf("config was not of expected type %q, but %q", config.AutomationType{}.ID(), c.Type.ID()))
+		return config.ResolvedEntity{}, errors.NewConfigDeployErr(c, fmt.Sprintf("config was not of expected type %q, but %q", config.AutomationType{}.ID(), c.Type.ID()))
 	}
 
 	var id string
@@ -54,24 +59,24 @@ func deployAutomation(ctx context.Context, client automationClient, properties p
 
 	resourceType, err := automationutils.ClientResourceTypeFromConfigType(t.Resource)
 	if err != nil {
-		return ResolvedEntity{}, newConfigDeployErr(c, fmt.Sprintf("failed to upsert automation object of type %s with id %s", t.Resource, id)).withError(err)
+		return config.ResolvedEntity{}, errors.NewConfigDeployErr(c, fmt.Sprintf("failed to upsert automation object of type %s with id %s", t.Resource, id)).WithError(err)
 	}
 
 	var resp *automation.Response
 	resp, err = client.Upsert(ctx, resourceType, id, []byte(renderedConfig))
 	if resp == nil || err != nil {
-		return ResolvedEntity{}, newConfigDeployErr(c, fmt.Sprintf("failed to upsert automation object of type %s with id %s", t.Resource, id)).withError(err)
+		return config.ResolvedEntity{}, errors.NewConfigDeployErr(c, fmt.Sprintf("failed to upsert automation object of type %s with id %s", t.Resource, id)).WithError(err)
 	}
 
 	name := fmt.Sprintf("[UNKNOWN NAME]%s", resp.ID)
-	if configName, err := extractConfigName(c, properties); err == nil {
+	if configName, err := resolve.ExtractConfigName(c, properties); err == nil {
 		name = configName
 	} else {
 		log.WithCtxFields(ctx).Warn("failed to extract name for automation object %q - ID will be used", resp.ID)
 	}
 
 	properties[config.IdParameter] = resp.ID
-	resolved := ResolvedEntity{
+	resolved := config.ResolvedEntity{
 		EntityName: name,
 		Coordinate: c.Coordinate,
 		Properties: properties,
